fix!: Change `ListCheckSuiteOptions.AppID` from `int` to `int64` (#3633)
BREAKING CHANGE: Change `ListCheckSuiteOptions.AppID` from `int` to `int64`

+++ github/checks_test.go
@@ -472,7 +472,7 @@ func TestChecksService_ListCheckSuiteForRef(t *testing.T) {
 
 	opt := &ListCheckSuiteOptions{
 		CheckName:   Ptr("testing"),
-		AppID:       Ptr(2),
+		AppID:       Ptr(int64(2)),
 		ListOptions: ListOptions{Page: 1},
 	}
 	ctx := context.Background()
